+++ packages/images/optimize.js
@@ -1,10 +1,15 @@
-const {extendDefaultPlugins} = require('svgo');
-
 function svgOptions() {
   return {
-    plugins: extendDefaultPlugins([{name: 'removeTitle'}]).filter(
-      (plugin) => plugin.name !== 'removeViewBox',
-    ),
+    plugins: [
+      {
+        name: 'preset-default',
+        params: {
+          overrides: {
+            removeViewBox: false,
+          },
+        },
+      },
+    ],
   };
 }